Move configuration into the config packge
We now have a `config.Config` struct which contains all the data in the
configuration file. Unfortunately, we still need a global config since
the routes do not have access to state. The `globalConfig` variable and
`config.Get{Field}` methods will be removed in a future path, which will
clean up a lot of testing infrastructure to not require writing changes
to disk.

The cleanup involved updating all the tests in `routes_tests.go`, so I
removed the setup/tearddown methods, which are not very idiomatic for go
unit tests and since you cannot defer into a parent scope it make sense
to have the following pattern:

```go
state := helpers.SetUp(t, ...)
defer helpers.TearDown(t, state)
```

for all the resources in a unit test.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/9771/

+++ server.go
@@ -10,7 +10,7 @@ import (
 const timeLayout = "02/Jan/2006:15:04:05 -0700"
 
 // Return a new RB Gateway HTTP server.
-func NewServer(port int) http.Server {
+func NewServer(port uint16) http.Server {
 	addr := fmt.Sprintf(":%d", port)
 
 	return http.Server{
